Enable CS expansion (fixes CaseConflict test)
